# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的记账系统 API，支持用户注册、登录、消费记录管理和数据导出功能",
        "title": "记账系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/income-categories": {
            "get": {
                "description": "获取所有收入类别列表，支持按名称模糊搜索",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-收入类别"
                ],
                "summary": "获取收入类别列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类别名称（模糊匹配）",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回类别列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的收入类别，支持设置名称、排序和颜色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-收入类别"
                ],
                "summary": "创建收入类别",
                "parameters": [
                    {
                        "description": "类别信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IncomeCategoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或类别名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/income-categories/{id}": {
            "put": {
                "description": "更新指定的收入类别信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-收入类别"
                ],
                "summary": "更新收入类别",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "类别ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的类别信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IncomeCategoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或类别名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "类别不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "软删除指定的收入类别",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-收入类别"
                ],
                "summary": "删除收入类别",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "类别ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "类别不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.IncomeCategoryCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "color": {
                    "description": "颜色代码，如 #10b981",
                    "type": "string",
                    "maxLength": 20
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "api.IncomeCategoryUpdateRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "maxLength": 20
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}